Deduplicate reload-loop tests into direct lifecycle tests and refresh orchestration doc
Remove the four orchestrator-mediated duplicates of the direct lifecycle
tests, move the loop-resilience and cancel-logging tests to
tests/test_config_lifecycle.py as direct tests, and update
docs/architecture/orchestration.md for the config_lifecycle/plugin_watch
split.
