+++ tests/test_config_reload.py
@@ -28,7 +28,7 @@ from mindroom.matrix.users import AgentMatrixUser
 from mindroom.message_target import MessageTarget
 from mindroom.orchestration.config_updates import ConfigUpdatePlan, _get_changed_agents, build_config_update_plan
 from mindroom.orchestration.plugin_watch import _drop_unconfigured_plugin_root_snapshots, watch_plugins_task
-from mindroom.orchestration.runtime import create_logged_task, log_startup_phase_finished, log_startup_phase_started
+from mindroom.orchestration.runtime import log_startup_phase_finished, log_startup_phase_started
 from mindroom.orchestrator import _MultiAgentOrchestrator, _watch_skills_task
 from mindroom.startup_errors import PermanentStartupError
 from mindroom.tool_system.plugins import PluginReloadResult
@@ -1427,188 +1427,6 @@ async def test_queued_config_reload_waits_for_in_flight_response_without_event_i
         await orchestrator.config_reload.cancel()
 
 
-@pytest.mark.asyncio
-async def test_queued_config_reload_surfaces_stuck_drain_and_forces_reload(
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
-) -> None:
-    """Queued reloads should warn and then force through a wedged drain."""
-    monkeypatch.setattr("mindroom.orchestration.config_lifecycle._CONFIG_RELOAD_DEBOUNCE_SECONDS", 0.01)
-    monkeypatch.setattr("mindroom.orchestration.config_lifecycle._CONFIG_RELOAD_IDLE_POLL_SECONDS", 0.01)
-    monkeypatch.setattr("mindroom.orchestration.config_lifecycle._CONFIG_RELOAD_DRAIN_WARNING_AFTER_SECONDS", 0.02)
-    monkeypatch.setattr("mindroom.orchestration.config_lifecycle._CONFIG_RELOAD_DRAIN_WARNING_INTERVAL_SECONDS", 1.0)
-    monkeypatch.setattr("mindroom.orchestration.config_lifecycle._CONFIG_RELOAD_DRAIN_FORCE_AFTER_SECONDS", 0.04)
-
-    logger_mock = MagicMock()
-    monkeypatch.setattr("mindroom.orchestration.config_lifecycle.logger", logger_mock)
-
-    orchestrator = _MultiAgentOrchestrator(runtime_paths=orchestrator_runtime_paths(tmp_path))
-    orchestrator.running = True
-
-    mock_bot = MagicMock(spec=AgentBot)
-    mock_bot.in_flight_response_count = 1
-    orchestrator.agent_bots["agent1"] = mock_bot
-    orchestrator.config_reload.update_config = AsyncMock(return_value=True)
-
-    orchestrator.config_reload.request_reload()
-    task = orchestrator.config_reload._reload_task
-    assert task is not None
-
-    await asyncio.wait_for(task, timeout=1)
-
-    orchestrator.config_reload.update_config.assert_awaited_once()
-    assert any(
-        call.args and call.args[0] == "Configuration reload still waiting for active responses to finish"
-        for call in logger_mock.warning.call_args_list
-    )
-    assert any(
-        call.args and call.args[0] == "Forcing configuration reload while responses are still active"
-        for call in logger_mock.error.call_args_list
-    )
-
-
-@pytest.mark.asyncio
-async def test_queued_config_reload_resets_drain_window_for_new_change(
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
-) -> None:
-    """A newer config change should get a fresh drain timeout window."""
-    monkeypatch.setattr("mindroom.orchestration.config_lifecycle._CONFIG_RELOAD_DEBOUNCE_SECONDS", 0.01)
-    monkeypatch.setattr("mindroom.orchestration.config_lifecycle._CONFIG_RELOAD_IDLE_POLL_SECONDS", 0.005)
-    monkeypatch.setattr("mindroom.orchestration.config_lifecycle._CONFIG_RELOAD_DRAIN_WARNING_AFTER_SECONDS", 1.0)
-    monkeypatch.setattr("mindroom.orchestration.config_lifecycle._CONFIG_RELOAD_DRAIN_WARNING_INTERVAL_SECONDS", 1.0)
-    monkeypatch.setattr("mindroom.orchestration.config_lifecycle._CONFIG_RELOAD_DRAIN_FORCE_AFTER_SECONDS", 0.12)
-
-    orchestrator = _MultiAgentOrchestrator(runtime_paths=orchestrator_runtime_paths(tmp_path))
-    orchestrator.running = True
-
-    mock_bot = MagicMock(spec=AgentBot)
-    mock_bot.in_flight_response_count = 1
-    orchestrator.agent_bots["agent1"] = mock_bot
-
-    loop = asyncio.get_running_loop()
-    started_at = loop.time()
-    update_called_at: float | None = None
-
-    async def fake_update_config() -> bool:
-        nonlocal update_called_at
-        update_called_at = loop.time()
-        return True
-
-    orchestrator.config_reload.update_config = AsyncMock(side_effect=fake_update_config)
-
-    orchestrator.config_reload.request_reload()
-    await asyncio.sleep(0.06)
-    orchestrator.config_reload.request_reload()
-
-    task = orchestrator.config_reload._reload_task
-    assert task is not None
-    await asyncio.wait_for(task, timeout=1)
-
-    assert update_called_at is not None
-    assert update_called_at - started_at >= 0.16
-    orchestrator.config_reload.update_config.assert_awaited_once()
-
-
-@pytest.mark.asyncio
-async def test_request_config_reload_ignores_changes_while_startup_is_in_progress(
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
-) -> None:
-    """Queued reloads should not start before the orchestrator is running."""
-    logger_mock = MagicMock()
-    monkeypatch.setattr("mindroom.orchestration.config_lifecycle.logger", logger_mock)
-
-    orchestrator = _MultiAgentOrchestrator(runtime_paths=orchestrator_runtime_paths(tmp_path))
-    orchestrator.config_reload.update_config = AsyncMock(return_value=True)
-
-    orchestrator.config_reload.request_reload()
-
-    assert orchestrator.config_reload._requested_at is None
-    assert orchestrator.config_reload._reload_task is None
-    orchestrator.config_reload.update_config.assert_not_awaited()
-    assert any(
-        call.args and call.args[0] == "Ignoring config change while startup is still in progress"
-        for call in logger_mock.info.call_args_list
-    )
-
-
-@pytest.mark.asyncio
-async def test_detached_task_cancel_logs_exception_instead_of_suppressing_silently(
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
-) -> None:
-    """Detached task cancellation should log unexpected failures and keep shutdown moving."""
-    logger_mock = MagicMock()
-    monkeypatch.setattr("mindroom.orchestration.runtime.logger", logger_mock)
-
-    orchestrator = _MultiAgentOrchestrator(runtime_paths=orchestrator_runtime_paths(tmp_path))
-    started = asyncio.Event()
-
-    async def fail_during_cancel() -> None:
-        started.set()
-        try:
-            await asyncio.Future()
-        except asyncio.CancelledError as err:
-            msg = "boom"
-            raise RuntimeError(msg) from err
-
-    orchestrator.config_reload._reload_task = create_logged_task(
-        fail_during_cancel(),
-        name="config_reload",
-        failure_message="config_reload failed",
-    )
-    await asyncio.wait_for(started.wait(), timeout=1)
-
-    await orchestrator.config_reload.cancel()
-
-    assert orchestrator.config_reload._reload_task is None
-    assert any(
-        call.args
-        and call.args[0] == "Detached task failed while being cancelled"
-        and call.kwargs.get("task_name") == "config_reload"
-        for call in logger_mock.debug.call_args_list
-    )
-    logger_mock.exception.assert_not_called()
-
-
-@pytest.mark.asyncio
-async def test_queued_config_reload_coalesces_rapid_changes(
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
-) -> None:
-    """Multiple quick config changes should produce one reload."""
-    monkeypatch.setattr("mindroom.orchestration.config_lifecycle._CONFIG_RELOAD_DEBOUNCE_SECONDS", 0.05)
-    monkeypatch.setattr("mindroom.orchestration.config_lifecycle._CONFIG_RELOAD_IDLE_POLL_SECONDS", 0.01)
-
-    orchestrator = _MultiAgentOrchestrator(runtime_paths=orchestrator_runtime_paths(tmp_path))
-    orchestrator.running = True
-
-    mock_bot = MagicMock(spec=AgentBot)
-    mock_bot.in_flight_response_count = 0
-    orchestrator.agent_bots["agent1"] = mock_bot
-
-    update_started = asyncio.Event()
-
-    async def fake_update_config() -> bool:
-        update_started.set()
-        return True
-
-    orchestrator.config_reload.update_config = AsyncMock(side_effect=fake_update_config)
-
-    orchestrator.config_reload.request_reload()
-    task = orchestrator.config_reload._reload_task
-    assert task is not None
-
-    await asyncio.sleep(0.02)
-    orchestrator.config_reload.request_reload()
-
-    await asyncio.wait_for(update_started.wait(), timeout=1)
-    await asyncio.wait_for(task, timeout=1)
-
-    orchestrator.config_reload.update_config.assert_awaited_once()
-
-
 def test_get_changed_agents_detects_culture_config_updates() -> None:
     """Agent restarts should trigger when their culture mode/assignment changes."""
     old_config = _runtime_bound_config(
@@ -2698,79 +2516,6 @@ async def test_run_cancellable_response_marks_thinking_placeholder_pending(
 
     assert captured_send["response_text"] == "Thinking..."
     assert captured_send["extra_content"] == {STREAM_STATUS_KEY: STREAM_STATUS_PENDING}
-
-
-@pytest.mark.asyncio
-async def test_failed_update_config_does_not_strand_queued_reload(
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
-) -> None:
-    """A failed update_config must not prevent a subsequently queued reload from running."""
-    monkeypatch.setattr("mindroom.orchestration.config_lifecycle._CONFIG_RELOAD_DEBOUNCE_SECONDS", 0.01)
-    monkeypatch.setattr("mindroom.orchestration.config_lifecycle._CONFIG_RELOAD_IDLE_POLL_SECONDS", 0.01)
-
-    orchestrator = _MultiAgentOrchestrator(runtime_paths=orchestrator_runtime_paths(tmp_path))
-    orchestrator.running = True
-
-    mock_bot = MagicMock(spec=AgentBot)
-    mock_bot.in_flight_response_count = 0
-    orchestrator.agent_bots["agent1"] = mock_bot
-
-    call_count = 0
-
-    async def failing_then_succeeding_update() -> bool:
-        nonlocal call_count
-        call_count += 1
-        if call_count == 1:
-            # First call fails; queue a new reload during the failure
-            orchestrator.config_reload.request_reload()
-            msg = "Simulated config update failure"
-            raise RuntimeError(msg)
-        return True
-
-    orchestrator.config_reload.update_config = AsyncMock(side_effect=failing_then_succeeding_update)
-    orchestrator.config_reload.request_reload()
-    task = orchestrator.config_reload._reload_task
-    assert task is not None
-
-    await asyncio.wait_for(task, timeout=2)
-
-    assert orchestrator.config_reload.update_config.await_count == 2
-
-
-@pytest.mark.asyncio
-async def test_config_change_during_update_config_triggers_second_reload(
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
-) -> None:
-    """A config change arriving while update_config runs should cause a second reload."""
-    monkeypatch.setattr("mindroom.orchestration.config_lifecycle._CONFIG_RELOAD_DEBOUNCE_SECONDS", 0.01)
-    monkeypatch.setattr("mindroom.orchestration.config_lifecycle._CONFIG_RELOAD_IDLE_POLL_SECONDS", 0.01)
-
-    orchestrator = _MultiAgentOrchestrator(runtime_paths=orchestrator_runtime_paths(tmp_path))
-    orchestrator.running = True
-
-    mock_bot = MagicMock(spec=AgentBot)
-    mock_bot.in_flight_response_count = 0
-    orchestrator.agent_bots["agent1"] = mock_bot
-
-    call_count = 0
-
-    async def update_config_with_second_change() -> bool:
-        nonlocal call_count
-        call_count += 1
-        if call_count == 1:
-            orchestrator.config_reload.request_reload()
-        return True
-
-    orchestrator.config_reload.update_config = AsyncMock(side_effect=update_config_with_second_change)
-    orchestrator.config_reload.request_reload()
-    task = orchestrator.config_reload._reload_task
-    assert task is not None
-
-    await asyncio.wait_for(task, timeout=2)
-
-    assert orchestrator.config_reload.update_config.await_count == 2
 
 
 @pytest.mark.asyncio
